+++ core/node/node_framework/src/implementations/layers/eth_sender/mod.rs
@@ -1,0 +1,4 @@
+pub mod aggregator;
+pub mod manager;
+
+pub use self::{aggregator::EthTxAggregatorLayer, manager::EthTxManagerLayer};
