+++ core/node/node_framework/src/implementations/resources/da_client.rs
@@ -3,7 +3,7 @@ use zksync_da_client::DataAvailabilityClient;
 use crate::resource::Resource;
 
 /// Represents a client of a certain DA solution.
-#[derive(Clone)]
+#[derive(Debug, Clone)]
 pub struct DAClientResource(pub Box<dyn DataAvailabilityClient>);
 
 impl Resource for DAClientResource {
